update deps and fix missing jwt token
