+++ handler/handler.go
@@ -1,11 +1,7 @@
 package handler
 
 import (
-	"fmt"
-
 	"github.com/jinzhu/gorm"
-	"github.com/labstack/echo"
-	"gopkg.in/go-playground/validator.v9"
 )
 
 type Handler struct {
@@ -16,30 +12,4 @@ func New(db *gorm.DB) *Handler {
 	return &Handler{
 		db: db,
 	}
-}
-
-type Error struct {
-	Errors map[string]interface{} `json:"errors"`
-}
-
-func NewError(err error) Error {
-	e := Error{}
-	e.Errors = make(map[string]interface{})
-	switch v := err.(type) {
-	case *echo.HTTPError:
-		e.Errors["message"] = v.Message
-	default:
-		e.Errors["body"] = v.Error()
-	}
-	return e
-}
-
-func NewValidatorError(err error) Error {
-	e := Error{}
-	e.Errors = make(map[string]interface{})
-	errs := err.(validator.ValidationErrors)
-	for _, v := range errs {
-		e.Errors[v.Field()] = fmt.Sprintf("%v", v.Tag())
-	}
-	return e
 }
