+++ handler/user.go
@@ -5,16 +5,17 @@ import (
 
 	"github.com/labstack/echo"
 	"github.com/xesina/golang-echo-realworld-example-app/models"
+	"github.com/xesina/golang-echo-realworld-example-app/utils"
 )
 
 func (h *Handler) Register(c echo.Context) error {
 	var u models.User
 	req := &userRegisterRequest{}
 	if err := req.bind(c, &u); err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	if err := h.db.Create(&u).Error; err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	return c.JSON(http.StatusCreated, newUserResponse(&u))
 }
@@ -23,15 +24,15 @@ func (h *Handler) Login(c echo.Context) error {
 	var u models.User
 	req := &userLoginRequest{}
 	if err := req.bind(c, &u); err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	if err := h.db.Where(&models.User{Email: u.Email}).First(&u).Error; err != nil {
 		//TODO: Return appropriate error
-		return c.JSON(http.StatusForbidden, NewError(err))
+		return c.JSON(http.StatusForbidden, utils.NewError(err))
 	}
 	if err := u.CheckPassword(req.User.Password); err != nil {
 		//TODO: Return appropriate error
-		return c.JSON(http.StatusForbidden, NewError(err))
+		return c.JSON(http.StatusForbidden, utils.NewError(err))
 	}
 	return c.JSON(http.StatusOK, newUserResponse(&u))
 }
@@ -39,7 +40,7 @@ func (h *Handler) Login(c echo.Context) error {
 func (h *Handler) CurrentUser(c echo.Context) error {
 	var u models.User
 	if err := h.db.First(&u, userIDFromToken(c)).Error; err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	return c.JSON(http.StatusOK, newUserResponse(&u))
 }
@@ -50,10 +51,10 @@ func (h *Handler) UpdateUser(c echo.Context) error {
 	req := newUserUpdateRequest()
 	req.populate(&u)
 	if err := req.bind(c, &u); err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	if err := h.db.Model(&u).Update(&u).Error; err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	return c.JSON(http.StatusOK, newUserResponse(&u))
 }
@@ -62,7 +63,7 @@ func (h *Handler) GetProfile(c echo.Context) error {
 	username := c.Param("username")
 	var u models.User
 	if err := h.db.Where(&models.User{Username: username}).Preload("Followers").First(&u).Error; err != nil {
-		return c.JSON(http.StatusNotFound, NewError(err))
+		return c.JSON(http.StatusNotFound, utils.NewError(err))
 	}
 	return c.JSON(http.StatusOK, newProfileResponse(c, &u))
 }
@@ -72,10 +73,10 @@ func (h *Handler) Follow(c echo.Context) error {
 	username := c.Param("username")
 	followerID := userIDFromToken(c)
 	if err := h.db.Where(&models.User{Username: username}).First(&u).Error; err != nil {
-		return c.JSON(http.StatusNotFound, NewError(err))
+		return c.JSON(http.StatusNotFound, utils.NewError(err))
 	}
 	if err := h.db.Model(&u).Association("Followers").Append(&models.Follow{FollowerID: followerID, FollowingID: u.ID}).Error; err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 	return c.JSON(http.StatusOK, newProfileResponse(c, &u))
 }
@@ -84,7 +85,7 @@ func (h *Handler) Unfollow(c echo.Context) error {
 	username := c.Param("username")
 	followerID := userIDFromToken(c)
 	if err := h.db.Where(&models.User{Username: username}).First(&u).Error; err != nil {
-		return c.JSON(http.StatusNotFound, NewError(err))
+		return c.JSON(http.StatusNotFound, utils.NewError(err))
 	}
 
 	f := models.Follow{
@@ -92,11 +93,11 @@ func (h *Handler) Unfollow(c echo.Context) error {
 		FollowingID: u.ID,
 	}
 	if err := h.db.Model(&u).Association("Followers").Find(&f).Error; err != nil {
-		return c.JSON(http.StatusNotFound, NewError(err))
+		return c.JSON(http.StatusNotFound, utils.NewError(err))
 	}
 
 	if err := h.db.Delete(f).Error; err != nil {
-		return c.JSON(http.StatusUnprocessableEntity, NewError(err))
+		return c.JSON(http.StatusUnprocessableEntity, utils.NewError(err))
 	}
 
 	return c.JSON(http.StatusOK, newProfileResponse(c, &u))
