+++ utils/errors.go
@@ -1,0 +1,33 @@
+package utils
+
+import (
+	"fmt"
+	"github.com/labstack/echo"
+	"gopkg.in/go-playground/validator.v9"
+)
+
+type Error struct {
+	Errors map[string]interface{} `json:"errors"`
+}
+
+func NewError(err error) Error {
+	e := Error{}
+	e.Errors = make(map[string]interface{})
+	switch v := err.(type) {
+	case *echo.HTTPError:
+		e.Errors["body"] = v.Message
+	default:
+		e.Errors["body"] = v.Error()
+	}
+	return e
+}
+
+func NewValidatorError(err error) Error {
+	e := Error{}
+	e.Errors = make(map[string]interface{})
+	errs := err.(validator.ValidationErrors)
+	for _, v := range errs {
+		e.Errors[v.Field()] = fmt.Sprintf("%v", v.Tag())
+	}
+	return e
+}
